Move base and (<~) to Situation.hs
