+++ src/Topics/StandardModernPrecision/TwoDiamondOpeners.hs
@@ -1,10 +1,10 @@
 module Topics.StandardModernPrecision.TwoDiamondOpeners(topic) where
 
 import Output(output)
-import Topic(Topic(..), base, (<~), wrap, Situations)
+import Topic(Topic(..), wrap, Situations)
 import Auction(forbid, pointRange, suitLength, minSuitLength, maxSuitLength,
                Action, alternatives, constrain, makePass, makeCall)
-import Situation(situation)
+import Situation(situation, base, (<~))
 import SituationHelpers(stdWrap, wrapVulDlr)
 import qualified Terminology as T
 import qualified CommonBids as B
